requirements.txt
# ============================================================
# 具身机器人仿真项目 - Python依赖
# 已在Python 3.9.6 (macOS arm64)上测试通过
# ============================================================

# 核心科学计算库
numpy==2.0.2
scipy==1.13.1

# 数据可视化
matplotlib==3.9.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
kiwisolver==1.4.7
pillow==11.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0

# 配置管理
PyYAML==6.0.3

# 物理仿真引擎 ✅
pybullet==3.2.7

# 工具库
packaging==26.0
six==1.17.0
zipp==3.23.0
importlib_resources==6.5.2

# ============================================================
# 安装说明
# ============================================================
# 
# 快速安装:
#   pip install -r requirements.txt
#
# 如果PyBullet安装失败，可尝试:
#   brew install cmake
#   pip install pybullet
#
# 验证安装:
#   python verify_setup.py
#
# ============================================================
